Reorganize agent behavior classes
