+++ CSharpSourceCode/Battle/AI/Components/WizardAIComponent.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using TaleWorlds.Library;
 using TaleWorlds.MountAndBlade;
+using TOW_Core.Battle.AI.AgentBehavior;
 using TOW_Core.Battle.AI.AgentBehavior.AgentCastingBehavior;
 using TOW_Core.Battle.AI.Decision;
 using TOW_Core.Utilities;
