+++ CSharpSourceCode/TOW_Core.csproj
@@ -324,19 +324,18 @@
     <Compile Include="Abilities\SpellBook\StatItemVM.cs" />
     <Compile Include="Abilities\SpellBook\SpellItemVM.cs" />
     <Compile Include="Abilities\SummonedAgentOrigin.cs" />
+    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehaviorConfiguration.cs" />
     <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\AbstractAgentCastingBehavior.cs" />
-    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\AgentCastingBehaviorConfiguration.cs" />
-    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\CenteredStaticAoEAgentCastingBehavior.cs" />
-    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\PreserveWindsAgentCastingBehavior.cs" />
+    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\AoEAdjacentCastingBehavior.cs" />
+    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\PreserveWindsCastingBehavior.cs" />
     <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\MissileCastingBehavior.cs" />
-    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\DirectionalMovingAoEAgentCastingBehavior.cs" />
+    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\AoEDirectionalCastingBehavior.cs" />
     <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\SummoningCastingBehavior.cs" />
-    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\TargetedStaticCastingBehavior.cs" />
-    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\TargetedStaticAoEAgentCastingBehavior.cs" />
+    <Compile Include="Battle\AI\AgentBehavior\AgentCastingBehavior\AoETargetedCastingBehavior.cs" />
     <Compile Include="Battle\AI\AgentBehavior\AgentTacticalBehavior\AbstractAgentTacticalBehavior.cs" />
-    <Compile Include="Battle\AI\AgentBehavior\AgentTacticalBehavior\AdjacentAoETacticalBehavior.cs" />
-    <Compile Include="Battle\AI\AgentBehavior\AgentTacticalBehavior\DirectionalAoETacticalBehavior.cs" />
-    <Compile Include="Battle\AI\AgentBehavior\AgentTacticalBehavior\KeepSafeAgentTacticalBehavior.cs" />
+    <Compile Include="Battle\AI\AgentBehavior\AgentTacticalBehavior\AoEAdjacentTacticalBehavior.cs" />
+    <Compile Include="Battle\AI\AgentBehavior\AgentTacticalBehavior\AoEDirectionalTacticalBehavior.cs" />
+    <Compile Include="Battle\AI\AgentBehavior\AgentTacticalBehavior\KeepSafeTacticalBehavior.cs" />
     <Compile Include="Battle\AI\Components\ArtilleryAI.cs" />
     <Compile Include="Battle\AI\Decision\Axis.cs" />
     <Compile Include="Battle\AI\Decision\BehaviorOption.cs" />
